Update Ruby version
Heroku deployments occasionally fail, such as this one:
https://dashboard.heroku.com/apps/travis-web-staging/activity/builds/6198e1c3-71f3-44cd-95e2-dc5c952faf06

The log says that “Only the most recent version of Ruby 2.1 is supported on Cedar-14”
which according to the linked document is currently 2.1.10.

It’s strange because Heroku deployments mostly succeed, so maybe Cedar 14 isn’t
always used? Regardless, it seems harmless to update to prevent these
intermittent failures.